Create dev utility class for some shared functionality
