+++ test/test-manifest.js
@@ -18,7 +18,7 @@
 const fs = require('fs');
 const path = require('path');
 const assert = require('assert');
-const {loadDefaultManifest, createManifestString} = require('../dev/build');
+const {getDefaultManifest, createManifestString} = require('../dev/yomichan-util');
 
 
 function loadManifestString() {
@@ -28,7 +28,7 @@ function loadManifestString() {
 
 function validateManifest() {
     const manifest1 = loadManifestString();
-    const manifest2 = createManifestString(loadDefaultManifest());
+    const manifest2 = createManifestString(getDefaultManifest());
     assert.strictEqual(manifest1, manifest2, 'Manifest data does not match.');
 }
